requirements.txt
Flask==3.0.2
flask-marshmallow==0.15.0
marshmallow==3.20.1
werkzeug>=3.0.1
python-dotenv==1.0.1
azure-monitor-opentelemetry==1.6.2
azure-monitor-opentelemetry-exporter==1.0.0b31
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-logging==0.48b0